requirements.txt
pandas
streamlit
requests
python-dotenv
groq
plotly
